backend-requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib
python-jose[cryptography]
pycryptodome
